Switch from tags to saves
No need to use a # before tags now. Need to check CPU and database usage, 'cause this will query the db for every word of every message in if you're reading this, you need to find something better to do and I know this is in the middle of a bloody sentence and I really wanna know who this bored fool is who's reading my horrible code and I swear that my Java is marginally better every group. Yes, IK that this is a really bad idea.

Also, I fear a huge spamwave, but all this bot does is spam, so, I guess that's fine. Oh, and save replies are autodeleted, so spammers, spam away! Just don't freeze the server plox.